fix: set workspace directory as safe in the entrypoint
`actions/checkout` already fixed the issue on their side in
[v3.0.1](https://github.com/actions/checkout/releases/tag/v3.0.1).
But since this is a container action, it can be run with another uid than the one
from `actions/checkout` and therefore we need to add the `safe.directory` config again.

Another way of fixing this would be to make this action run with the same uid as the github runner,
 but as mentioned [here](https://github.com/actions/checkout/issues/760#issuecomment-1098135955)
we're not sure whether the `1000` uid of the runner is reliable.

The different `uid` issue seems more like a Github Actions ecosystem problem,
maybe they'll take care of it later.